refactoring of apiclient resource
